test: slice tests moved to slices folder; added test prepare a ingredient;
